voice-coche-api/requirements.txt
click==8.1.3
Flask==3.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask_Login==0.6.3
Jinja2==3.1.2
pillow==10.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
Werkzeug==3.0.1
cryptography==42.0.5
APScheduler==3.10.4